move assessment table in class layout by id to assessment tab, fix clicking row on assessment table will take user to note instead of clicking assessment name
